Move more code from render.ts to image.ts and new api.ts
