+++ src/index.ts
@@ -15,5 +15,7 @@ export {
   getSlices,
   getMinMaxValues,
 } from "./utils";
-export { renderThumbnail, renderImage } from "./render";
+export { renderImage } from "./render";
 export { LUTS, getLuts } from "./luts";
+export { NgffImage } from "./image";
+export { renderThumbnail } from "./api";
